requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.1
pydantic==2.5.0
feishu-python-sdk==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
cryptography==41.0.7
redis==5.0.1
celery==5.3.4